New Searchview style, not in the toolbar Bug Fix wrong data type main.pressure in WeatherResponse

+++ app/src/main/java/com/openweather/challenge/openweatherapp/model/Main.java
@@ -13,7 +13,7 @@ package com.openweather.challenge.openweatherapp.model;
  */
 public class Main {
     public double temp;
-    public int pressure;
+    public double pressure;
     public int humidity;
     public double temp_min;
     public double temp_max;
